WritingTutor_with_Chatomatic-master/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.12
async-timeout==4.0.2
attrs==22.2.0
awscli==1.18.163
Babel==2.8.0
bleach==3.1.5
blis==0.7.9
botocore==1.19.3
catalogue==2.0.8
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.1.0
ChatterBot==0.8.7
chatterbot-corpus==1.1.4
click==7.1.2
colorama==0.4.3
confection==0.0.4
coverage==5.2
coveralls==2.1.1
cymem==2.0.3
DateTime==4.3
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.5.0/de_core_news_sm-3.5.0-py3-none-any.whl
docopt==0.6.2
docutils==0.15.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
filelock==3.9.0
flake8==3.8.3
Flask==1.0.4
Flask-Cors==3.0.10
frozenlist==1.3.3
future==0.18.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.12.1
hyperframe==5.2.0
idna==2.10
imagesize==1.2.0
install==1.3.5
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.2.0
keyring==21.2.1
langcodes==3.3.0
MarkupSafe==1.1.1
mathparse==0.1.2
mccabe==0.6.1
multidict==6.0.4
murmurhash==1.0.2
nltk==3.5
nose==1.3.7
numpy==1.19.0
oauthlib==3.1.0
openai==0.27.2
packaging==20.4
pathy==0.10.1
Pillow==7.2.0
Pint==0.14
pkginfo==1.5.0.1
plac==1.1.3
preshed==3.0.2
pyasn1==0.4.8
pycodestyle==2.6.0
pydantic==1.10.5
pyflakes==2.2.0
Pygments==2.6.1
pymongo==3.10.1
pyparsing==2.4.7
pyperclip==1.8.0
python-dateutil==2.6.1
python-twitter==3.5
pytz==2020.1
PyYAML==6.0
rank-bm25==0.2.2
readme-renderer==26.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.5
s3transfer==0.3.3
scikit-learn==1.2.1
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.15.0
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.0.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacytextblob==4.0.0
Sphinx==3.0.4
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.2.19
srsly==2.4.6
textblob==0.15.3
textblob-de==0.4.3
thinc==8.1.8
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.48.0
transformers==4.26.1
twine==3.2.0
twython==3.8.2
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.25.9
wasabi==1.1.1
webencodings==0.5.1
Werkzeug==1.0.1
yarl==1.8.2
zope.interface==5.1.2
